# docs/swagger.yaml
basePath: /api/v1
definitions:
  helpers.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.LoginUser:
    properties:
      email:
        example: iqbal@mail.com
        type: string
      password:
        example: iqbal123
        type: string
    required:
    - email
    - password
    type: object
  models.RegisterUser:
    properties:
      age:
        example: 22
        type: integer
      email:
        example: iqbal@mail.com
        type: string
      password:
        example: iqbal123
        type: string
      username:
        example: iqbaludinm
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
...
host: localhost:8080
info:
  contact:
    email: iqbaludinm14@gmail.com
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample service for social-media.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MyGram API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in the user with the provided email and password.
      parameters:
      - description: Credential for Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: User Login to MyGram
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register user in MyGram Service
      parameters:
      - description: Credential for Register
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.RegisterUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/helpers.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helpers.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helpers.Response'
      summary: User Register to MyGram
      tags:
      - Auth
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"